requirements.txt
Pillow>=9.0.0
Flask>=2.3.0
Flask-SocketIO>=5.3.0
gevent>=23.0.0
gevent-websocket>=0.10.1
python-socketio>=5.9.0
gunicorn>=21.0.0
